+++ app/src/main/java/me/saket/dank/ui/submission/SubmissionImageLoader.java
@@ -117,9 +117,12 @@ public class SubmissionImageLoader {
     } else {
       // Images supplied by Reddit are static, so cannot optimize for GIFs.
       String defaultImageUrl = mediaLink.lowQualityUrl();
-      return mediaLink.isGif()
-          ? defaultImageUrl
-          : ImageWithMultipleVariants.Companion.of(redditPreviews).findNearestFor(deviceDisplaySize.getWidth(), defaultImageUrl);
+      return mediaLink.isGif() ? defaultImageUrl :
+          ImageWithMultipleVariants.Companion.of(redditPreviews).findNearestFor(
+              deviceDisplaySize.getWidth(),
+              ImageWithMultipleVariants.DEFAULT_VIEWER_MIN_WIDTH,
+              defaultImageUrl
+          );
     }
   }