+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 
 - Show smaller Imgur images when data saving mode enabled ([#256](https://github.com/Tunous/Dawn/pull/256))
+- Don't show too small reddit previews ([#264](https://github.com/Tunous/Dawn/pull/264))
 
 ## [0.9.2] - 2020-06-13